changed controller to return xMin etc. as doubles

+++ src/controller/Controller.java
@@ -33,7 +33,7 @@ public final class Controller {
   private static FileLoader fileLoader;
   private Connection[] connections;
   private ConnectionQuadTree qt;
-	private static int xMin, yMin, xMax, yMax;
+	private static double xMin, yMin, xMax, yMax;
 
 /**
    * Constructor for this class loads connections and points from FileLoader
@@ -46,10 +46,10 @@ public final class Controller {
     try {
       fileLoader = new FileLoaderConnectionOnly();
       connections = fileLoader.getConnections();
-      xMin = fileLoader.getxMin().intValue();
-      yMin = fileLoader.getyMin().intValue();
-      xMax = fileLoader.getxMax().intValue();
-      yMax = fileLoader.getyMax().intValue();
+      xMin = fileLoader.getxMin().doubleValue();
+      yMin = fileLoader.getyMin().doubleValue();
+      xMax = fileLoader.getxMax().doubleValue();
+      yMax = fileLoader.getyMax().doubleValue();
       qt = fileLoader.getConnectionQuadTree();
     } catch (IOException e) {
       System.out.println("Fileloader: " + e);
@@ -141,19 +141,19 @@ public final class Controller {
     map.showAll();
   }
 
-  public static int getxMin() {
+  public static double getxMin() {
 		return xMin;
 	}
 
-	public static int getyMin() {
+	public static double getyMin() {
 		return yMin;
 	}
 
-	public static int getxMax() {
+	public static double getxMax() {
 		return xMax;
 	}
 
-	public static int getyMax() {
+	public static double getyMax() {
 		return yMax;
 	}
   /**
